# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agro API",
        "description": "Documentación de la API de autenticación con Swagger en Laravel 11",
        "contact": {
            "email": "soporte@tuempresa.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Registra un nuevo usuario",
                "operationId": "b9e071918fa6250b23cdeddd7643e0a1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "email",
                                    "role",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "usuario123"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "usuario@example.com"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "user"
                                        ],
                                        "example": "user"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "securePassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado correctamente"
                    },
                    "422": {
                        "description": "Error en validación de datos"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Inicia sesión y obtiene un token JWT",
                "operationId": "d981c87d4b23485ce8fa578d2d9f95f6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "usuario123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "securePassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Autenticación exitosa, devuelve el token JWT"
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Obtiene los datos del usuario autenticado",
                "operationId": "7f111548d336760611bdb9d837cc332e",
                "responses": {
                    "200": {
                        "description": "Datos del usuario autenticado"
                    },
                    "401": {
                        "description": "Token inválido o expirado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Cierra sesión e invalida el token JWT",
                "operationId": "e44cd8b1e02c80892153cba48f159c36",
                "responses": {
                    "200": {
                        "description": "Cierre de sesión exitoso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/send-email-password-change": {
            "post": {
                "tags": [
                    "ChangePassword"
                ],
                "summary": "Envía un código de verificación para el cambio de contraseña",
                "description": "Genera un código de verificación que se enviará al correo electrónico del usuario para realizar el cambio de contraseña.",
                "operationId": "sendEmailPasswordChange",
                "requestBody": {
                    "description": "Los datos necesarios para el cambio de contraseña",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "usuario@ejemplo.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Código de verificación enviado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "Se envio el codigo de verificación revise su correo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos de entrada inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El campo email es obligatorio y debe ser una dirección de correo válida"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/change-password": {
            "post": {
                "tags": [
                    "ChangePassword"
                ],
                "summary": "Cambia la contraseña del usuario utilizando un código de verificación",
                "description": "Permite al usuario cambiar su contraseña mediante un código de verificación enviado por correo electrónico.",
                "operationId": "changePassword",
                "requestBody": {
                    "description": "Datos necesarios para cambiar la contraseña",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "code",
                                    "password_new"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "usuario@ejemplo.com"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "password_new": {
                                        "description": "Contraseña nueva con al menos 8 caracteres",
                                        "type": "string",
                                        "example": "NuevaContraseña123"
                                    },
                                    "password_new_confirmation": {
                                        "description": "Confirmar contraseña nueva con al menos 8 caracteres",
                                        "type": "string",
                                        "example": "NuevaContraseña123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contraseña cambiada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "Su contraseña se cambio"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos de entrada inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "La contraseña debe tener al menos 8 caracteres"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado o código de verificación incorrecto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Credenciales inválidas"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "410": {
                        "description": "Código de verificación expirado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El código de verificación ha expirado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}